First pass style improvements
