+++ perda/core_data_structures/CLAUDE.md
@@ -1,0 +1,5 @@
+`core_data_structures` is a submodule focused on representing and manipulating the fundamental time-series variables in our log files.
+
+- Whenever possible, separate data representation from data manipulation.
+
+- Keep files small, descriptively-named, and focused on well-defined and narrrowly-scoped purposes.
