+++ perda/core_data_structures/__init__.py
@@ -1,6 +1,7 @@
 from .data_instance import *
 from .deduplication import *
 from .joins import *
-from .resampling_helpers import *
+from .masking import *
+from .resampling import *
 from .single_run_data import *
 from .split_helpers import *
