+++ perda/utils/__init__.py
@@ -1,8 +1,8 @@
 from ..units import *
 from .data_summary import *
-from .filtering import *
 from .frequency_analysis import *
 from .gps_analysis import *
 from .integrate import *
 from .preprocessing import *
-from .resampling import *
+from .signal_processing import *
+from .time_alignment import *
