+++ perda/utils/signal_processing.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import numpy as np
 from numpy import float64
 from numpy.typing import NDArray
